refactor: components and pages location
