--- src/pages/about-us/ui/desktop/index.ts
@@ -1,1 +1,0 @@
-export { default as AboutUsPage } from './AboutUsPage.vue'
