--- src/pages/contacts/ui/desktop/index.ts
@@ -1,1 +1,0 @@
-export { default as ContactsPage } from './ContactsPage.vue'
