+++ src/pages/gallery/ui/GalleryPage.vue
