--- src/pages/gallery/ui/desktop/index.ts
@@ -1,1 +1,0 @@
-export { default as GalleryPage } from './GalleryPage.vue'
