+++ src/pages/gallery/ui/index.ts
@@ -1,1 +1,1 @@
-export { GalleryPage as GalleryDesktop } from './desktop'
+export { default as GalleryPage } from './GalleryPage.vue'
