+++ src/pages/home/ui/components/GalleryBlock.vue
