+++ src/pages/home/ui/components/ImageCarousel.vue
