+++ src/pages/home/ui/components/OurPartners.vue
@@ -22,5 +22,3 @@
     </div>
   </div>
 </template>
-
-<script setup lang="ts"></script>
