+++ src/pages/home/ui/components/ShopContacts.vue
