--- src/pages/home/ui/desktop/index.ts
@@ -1,1 +1,0 @@
-export { default as HomePage } from './HomePage.vue'
