+++ src/pages/index.ts
@@ -1,6 +1,6 @@
-export { ContactsDesktop } from './contacts'
-export { HomeDesktop } from './home'
-export { AboutUsDesktop } from './about-us'
-export { NewsDetailsDesktop, NewsListDesktop, NewsMainDesktop } from './news'
-export { NotFoundDesktop } from './not-found'
-export { GalleryDesktop } from './gallery'
+export { ContactsPage } from './contacts'
+export { HomePage } from './home'
+export { AboutUsPage } from './about-us'
+export { NewsDetailsPage, NewsListPage, NewsMainPage } from './news'
+export { NotFoundPage } from './not-found'
+export { GalleryPage } from './gallery'
