+++ src/pages/news/ui/NewsDetailsPage.vue
