+++ src/pages/news/ui/NewsMainPage.vue
