+++ src/pages/news/ui/components/NewsArticle.vue
@@ -14,7 +14,7 @@
 <script setup lang="ts">
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
-import type { NewsDetails } from '../../../model'
+import type { NewsDetails } from '../../model'
 
 const { locale } = useI18n()
