+++ src/pages/news/ui/components/NewsCard.vue
@@ -25,7 +25,7 @@
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { PageLink } from '@/shared'
-import type { NewsDetails } from '../../../model'
+import type { NewsDetails } from '../../model'
 
 const { locale } = useI18n()
