--- src/pages/news/ui/desktop/index.ts
@@ -1,5 +1,0 @@
-import NewsDetailsPage from './NewsDetailsPage.vue'
-import NewsListPage from './NewsListPage.vue'
-import NewsMainPage from './NewsMainPage.vue'
-
-export { NewsDetailsPage, NewsListPage, NewsMainPage }
