+++ src/pages/news/ui/index.ts
@@ -1,5 +1,3 @@
-export {
-  NewsListPage as NewsListDesktop,
-  NewsDetailsPage as NewsDetailsDesktop,
-  NewsMainPage as NewsMainDesktop
-} from './desktop'
+export { default as NewsListPage } from './NewsListPage.vue'
+export { default as NewsDetailsPage } from './NewsDetailsPage.vue'
+export { default as NewsMainPage } from './NewsMainPage.vue'
