+++ src/pages/not-found/ui/NotFoundPage.vue
