--- src/pages/not-found/ui/desktop/index.ts
@@ -1,1 +1,0 @@
-export { default as NotFoundPage } from './NotFoundPage.vue'
